# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This microservice has the responsability of creating and validating QR Codes for a Queue",
        "title": "QR Code Microservice",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "http://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/qrcode": {
            "get": {
                "description": "Returns QR Code for a Giver",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "qrcode"
                ],
                "summary": "Returns QR Code for a Giver",
                "responses": {
                    "200": {
                        "description": "QR Code for Queue",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/ticket": {
            "get": {
                "description": "Returns ticket number for the queue",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "qrcode",
                    "ticket"
                ],
                "summary": "Returns ticket number for the queue",
                "responses": {
                    "200": {
                        "description": "Ticket number for queue",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "gin.H": {
            "type": "object",
            "additionalProperties": {}
        }
    }
}